% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFeatures.R
\name{computeFeatures}
\alias{computeFeatures}
\title{compute the nn-features used to detected gridded datasets for all datasets in a file}
\usage{
computeFeatures(file, saveDir = "C:/Users/ftw712/Desktop/")
}
\arguments{
\item{file}{a file with occurrence points probably named "uniquePointsDatasetkey.csv"}

\item{saveDir}{where to save where to save griddedDatasets.rda}
}
\value{
nothing. Saves griddedDatasets.rda in SaveDir.
}
\description{
compute the nn-features used to detected gridded datasets for all datasets in a file
}
