% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGriddedKeys.R
\name{getGriddedKeys}
\alias{getGriddedKeys}
\title{Get GBIF datasetkeys that are rasterized/gridded/equally-spaced}
\usage{
getGriddedKeys(minDistance = 0, minCount = 0, minPercent = 0,
  limit = 1000)
}
\arguments{
\item{minDistance}{Filters by the minimum distance in decimal degrees of the most common nn-distance. Default is 0.}

\item{minCount}{Filters by the minimum number of points that have the same nn-distance. Default is 0.}

\item{minPercent}{Filters by the  minimum percentage of occurrence points with the same nn-distance. Default is 0.}

\item{limit}{the maximum number of datasetkeys to return. Default is 1000. There are only around 600 gridded datasets on GBIF.org.}
}
\value{
Returns a vector of uuid datasetkeys
}
\description{
Gets all the gridded uuid datasetkeys from the GBIF.org registry that pass certain filters.
}
\examples{

\dontrun{

library(rgbif)
library(dplyr)
# taxon key for Syntrichia norvegica Weber, 1804 (a moss)
taxonKey=2671266
# Get occurrence data for a moss occurring in SWEDEN
occData <- occ_data(taxonKey=taxonKey,
                   country="SE", # get only from SWEDEN to limit results
                   hasGeospatialIssue=FALSE, # remove those with other geospatial issues
                   hasCoordinate=TRUE, # get only with coordinates
                   limit=1000)$data # get first 1000 records
# only get dataset keys with spacing greater than around 10 km
keysToFilter <- gbifgridded::getGriddedKeys(minDistance = 0.1)
nrow(occData)
cleanData <- occData \%>\% # should be around 140 occurrence records
 filter(!datasetKey \%in\% keysToFilter)
nrow(cleanData) # should remove around 50 low resolution occurrences

}

}
\references{
\url{https://data-blog.gbif.org/post/finding-gridded-datasets/}
}
