% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNNFeature.R
\name{getNNFeature}
\alias{getNNFeature}
\title{Get the nn-features for gridded dataset classification}
\usage{
getNNFeature(D, key, k = 4)
}
\arguments{
\item{D}{data.frame with named columns of decimalLongitude and decimalLatitude}

\item{key}{a character string of a uuid of a dataset}

\item{k}{do not change from 4}
}
\value{
Returns data.frame of results.
}
\description{
Get the nn-features for gridded dataset classification
}
\examples{

\dontrun{



}

}
