% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{griddedDatasets}
\alias{griddedDatasets}
\title{griddedDatasets}
\format{A data.frame 11817 obs. of  3 variables :
\describe{
  \item{datasetkey}{uuid of dataset}
  \item{data_resource_id}{the id of dataset used prior to 2013-12-20}
  \item{created}{date the dataset was created}
  ...
}}
\source{
\url{}
}
\usage{
griddedDatasets
}
\description{
griddedDatasets
}
\keyword{datasets}
