% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFeatures.R
\name{computeFeatures}
\alias{computeFeatures}
\title{Add taxon info to a data.frame.}
\usage{
computeFeatures(file, saveDir = "C:/Users/ftw712/Desktop/")
}
\arguments{
\item{.data}{data.frame to be passed in probably through a pipe %>%}

\item{key}{unquoted column name for taxonkeys to be passed in}
}
\value{
Returns data.frame of results.
}
\description{
Add taxon info to a data.frame.
}
\examples{

\dontrun{

D \%>\% addTaxonInfo(specieskey)

}

}
