% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNNFeature.R
\name{getNNFeature}
\alias{getNNFeature}
\title{Add taxon info to a data.frame.}
\usage{
getNNFeature(D, key, k)
}
\arguments{
\item{key}{unquoted column name for taxonkeys to be passed in}

\item{.data}{data.frame to be passed in probably through a pipe %>%}
}
\value{
Returns data.frame of results.
}
\description{
Add taxon info to a data.frame.
}
\examples{

\dontrun{

D \%>\% addTaxonInfo(specieskey)

}

}
